fix(phpstan): strict-rules 开启全量清零 420→0，修复 14 处运行时回归，Pint 合规
- 引入 phpstan/phpstan-strict-rules ^2.0 并在 phpstan.neon 启用；仅裁剪
  dynamicCallOnStaticMethod（Larastan 将 Eloquent 静态 Builder/validate 宏
  误判为动态静态调用）与 nullsafe.neverNull 两条框架层规则，均附理由注释；
  baseline 保持为空
- empty()/array_filter/in_array/布尔条件三大批次恒等变换：=== null/''/0/[]、
  isset()、(bool) 包装、preg_match !== 1、filter_var === false、is_numeric 恒等化
- 测试基线对比（stash 原始代码 608 全绿）甄别并修复 14 处变换回归，根因均为
  "phpstan 静态类型 ≠ Eloquent cast/DB 可空列运行时真相"：
  - AdminUserUpdate / User::getPlanSettings：cast 'array' 对 null 列值返回 null，
    array_merge 前需显式窄化
  - Code::redemptionIssue/recordRedemption：max_redemptions 可空（null/0=不限），
    !== 0 变换丢 null 语义导致新码误判 fully_redeemed
- 206 个改动文件经 Pint（laravel preset）全量合规

验证：phpstan level 10 + strict-rules 0 错误；phpunit 608 tests / 2328 assertions 全绿

+++ app/Http/Controllers/Api/v1/AnalyticsController.php
@@ -337,7 +337,7 @@ class AnalyticsController extends Controller
      */
     protected function authorizeWebsite(Website $website): void
     {
-        if ((int) $website->user_id !== (int) Auth::id() && ! $this->user()->isAdmin()) {
+        if ($website->user_id !== (int) Auth::id() && ! $this->user()->isAdmin()) {
             abort(403, 'Unauthorized');
         }
     }
