+++ app/Http/Controllers/Api/v1/WebsiteController.php
@@ -91,7 +91,7 @@ class WebsiteController extends Controller
      */
     protected function authorizeWebsite(Website $website): void
     {
-        if ((int) $website->user_id !== (int) Auth::id() && ! $this->user()->isAdmin()) {
+        if ($website->user_id !== (int) Auth::id() && ! $this->user()->isAdmin()) {
             abort(403, 'Unauthorized');
         }
     }
